Now use std::cerr for error instead of cout

+++ src/validate.cpp
@@ -21,7 +21,7 @@ bool equalsIgnoreCase(const std::string& a, const std::string& b)
 {
   if(a.size() != b.size())
     return false;
-  for (int i = 0; i < a.size(); i++)
+  for (size_t i = 0; i < a.size(); i++)
     if(std::tolower(a[i]) != std::tolower(b[i]))
       return false;
   return true;
